Bullet Native: - some fixes in SixDof and Slider joints

+++ branches/bullet-native/native/com_jme3_bullet_joints_SixDofJoint.cpp
@@ -48,7 +48,7 @@ extern "C" {
     JNIEXPORT jlong JNICALL Java_com_jme3_bullet_joints_SixDofJoint_getRotationalLimitMotor
     (JNIEnv * env, jobject object, jlong jointId, jint index) {
         btGeneric6DofConstraint* joint = (btGeneric6DofConstraint*) jointId;
-        return (long) joint->getRotationalLimitMotor(index);
+        return joint->getRotationalLimitMotor(index);
     }
 
     /*
@@ -59,7 +59,7 @@ extern "C" {
     JNIEXPORT jlong JNICALL Java_com_jme3_bullet_joints_SixDofJoint_getTranslationalLimitMotor
     (JNIEnv * env, jobject object, jlong jointId) {
         btGeneric6DofConstraint* joint = (btGeneric6DofConstraint*) jointId;
-        return (long) joint->getTranslationalLimitMotor();
+        return joint->getTranslationalLimitMotor();
     }
 
     /*
@@ -132,7 +132,7 @@ extern "C" {
         jmeBulletUtil::convert(env, pivotB, &transB->getOrigin());
         jmeBulletUtil::convert(env, rotB, &transB->getBasis());
         btGeneric6DofConstraint* joint = new btGeneric6DofConstraint(*bodyA, *bodyB, *transA, *transB, useLinearReferenceFrameA);
-        return (long) joint;
+        return (long)joint;
     }
 #ifdef __cplusplus
 }
